% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwot.R
\name{umap}
\alias{umap}
\title{Dimensionality Reduction with UMAP}
\usage{
umap(X, n_neighbors = 15, n_components = 2, n_epochs = NULL, alpha = 1,
  init = "spectral", spread = 1, min_dist = 0.01, set_op_mix_ratio = 1,
  local_connectivity = 1, bandwidth = 1, gamma = 1,
  negative_sample_rate = 5, a = NULL, b = NULL, nn_method = NULL,
  n_trees = 50, search_k = 2 * n_neighbors * n_trees, approx_pow = FALSE,
  n_threads = RcppParallel::defaultNumThreads()/2, grain_size = 1000,
  verbose = getOption("verbose", TRUE))
}
\arguments{
\item{X}{Input data. Can be a \code{\link{data.frame}}, \code{\link{matrix}},
\code{\link[stats]{dist}} object or \code{\link[Matrix]{sparseMatrix}}.
A sparse matrix is interpreted as a distance matrix and both implicit and
explicit zero entries are ignored. Set zero distances you want to keep to
an arbitrarily small non-zero value (e.g. \code{1e-10}). Matrix and data
frames should contain one observation per row. Data frames will have any
non-numeric columns removed.}

\item{n_neighbors}{The size of local neighborhood (in terms of number of
neighboring sample points) used for manifold approximation. Larger values
result in more global views of the manifold, while smaller values result in
more local data being preserved. In general values should be in the range
\code{2} to \code{100}.}

\item{n_components}{The dimension of the space to embed into. This defaults
to \code{2} to provide easy visualization, but can reasonably be set to any
integer value in the range \code{2} to \code{100}.}

\item{n_epochs}{Number of epochs to use during the optimization of the
embedded coordinates. By default, this value is set to \code{500} for datasets
containing 10,000 vertices or less, and \code{200} otherwise.}

\item{alpha}{Initial learning rate used in optimization of the coordinates.}

\item{init}{Type of initialization for the coordinates. Options are:
\itemize{
  \item \code{"spectral"} Spectral embedding using the normalized Laplacian
  of the fuzzy 1-skeleton, with Gaussian noise added.
  \item \code{"normlaplacian"}. Spectral embedding using the normalized
  Laplacian of the fuzzy 1-skeleton, without noise.
  \item \code{"random"}. Coordinates assigned using a uniform random
  distribution between -10 and 10.
  \item \code{"lvrandom"}. Coordinates assigned using a Gaussian
  distribution with standard deviation 1e-4, as in LargeVis
  (Tang et al., 2016).
  \item \code{"laplacian"}. Spectral embedding using the Laplacian Eigenmap
  (Belkin and Niyogi, 2002).
  \item \code{"spca"}. The first two principal components from PCA of
  \code{X} if \code{X} is a data frame, and from a 2-dimensional classical
  MDS if \code{X} is of class \code{"dist"}. These vectors are then scaled
  so their standard deviation is 0.0001.
  \item A matrix of initial coordinates.
}}

\item{spread}{The effective scale of embedded points. In combination with
\code{min_dist}, this determines how clustered/clumped the embedded points
are.}

\item{min_dist}{The effective minimum distance between embedded points.
Smaller values will result in a more clustered/clumped embedding where
nearby points on the manifold are drawn closer together, while larger
values will result on a more even dispersal of points. The value should be
set relative to the \code{spread} value, which determines the scale at
which embedded points will be spread out.}

\item{set_op_mix_ratio}{Interpolate between (fuzzy) union and intersection as
the set operation used to combine local fuzzy simplicial sets to obtain a
global fuzzy simplicial sets. Both fuzzy set operations use the product
t-norm. The value of this parameter should be between \code{0.0} and
\code{1.0}; a value of \code{1.0} will use a pure fuzzy union, while
\code{0.0} will use a pure fuzzy intersection.}

\item{local_connectivity}{The local connectivity required -- i.e. the number
of nearest neighbors that should be assumed to be connected at a local
level. The higher this value the more connected the manifold becomes
locally. In practice this should be not more than the local intrinsic
dimension of the manifold.}

\item{bandwidth}{The effective bandwidth of the kernel if we view the
algorithm as similar to Laplacian eigenmaps. Larger values induce more
connectivity and a more global view of the data, smaller values concentrate
more locally.}

\item{gamma}{Weighting applied to negative samples in low dimensional embedding
optimization. Values higher than one will result in greater weight
being given to negative samples.}

\item{negative_sample_rate}{The number of negative edge/1-simplex samples to
use per positive edge/1-simplex sample in optimizing the low dimensional
embedding.}

\item{a}{More specific parameters controlling the embedding. If \code{NULL}
these values are set automatically as determined by \code{min_dist} and
\code{spread}.}

\item{b}{More specific parameters controlling the embedding. If \code{NULL}
these values are set automatically as determined by \code{min_dist} and
\code{spread}.}

\item{nn_method}{Method for finding nearest neighbors. Options are:
\itemize{
  \item \code{"fnn"}. Use exact nearest neighbors via the
    \href{https://cran.r-project.org/package=FNN}{FNN} package.
  \item \code{"annoy"} Use approximate nearest neighbors via the
    \href{https://cran.r-project.org/package=RcppAnnoy}{RcppAnnoy} package.
 }
By default, if \code{X} has less than 4,096 vertices, the exact nearest
neighbors are found. Otherwise, approximate nearest neighbors are used.}

\item{n_trees}{Number of trees to build when constructing the nearest
neighbor index. The more trees specified, the larger the index, but the
better the results. With \code{search_k}, determines the accuracy of the
ANNOY nearest neighbor search. Only used if the \code{nn_method} is
\code{"annoy"}. Sensible values are between \code{10} to \code{100}.}

\item{search_k}{Number of nodes to search during the neighbor retrieval. The
larger k, the more the accurate results, but the longer the search takes.
With \code{n_trees}, determines the accuracy of the ANNOY nearest neighbor
search. Only used if the \code{nn_method} is \code{"annoy"}.}

\item{approx_pow}{If \code{TRUE}, use an approximation to the power function
in the UMAP gradient, from
\url{https://martin.ankerl.com/2012/01/25/optimized-approximative-pow-in-c-and-cpp/}.}

\item{n_threads}{Number of threads to use during the nearest neighbor search.
Default is half of \code{\link[RcppParallel]{defaultNumThreads}}. Only
applies if \code{nn_method = "annoy"}.}

\item{grain_size}{Minimum batch size for multithreading. If the number of
items to process in a thread falls below this number, then no threads will
be used. Used in conjunction with \code{n_threads}. Only applies during
the nearest neighbor search and if \code{nn_method = "annoy"}.}

\item{verbose}{If \code{TRUE}, log details to the console.}
}
\value{
A matrix of optimized coordinates.
}
\description{
Carry out dimensionality reduction of a dataset using the Uniform Manifold
Approximation and Projection (UMAP) method (McInnes & Healy, 2018). Some of
the following help text is lifted verbatim from the Python reference
implementation at \url{https://github.com/lmcinnes/umap}.
}
\examples{
\dontrun{
iris_umap <- umap(iris, n_neighbors = 50, alpha = 0.5, init = "random")

# Load mnist from somewhere, e.g.
# devtools::install_github("jlmelville/snedata")
# mnist <- snedata::download_mnist()
mnist_umap <- umap(mnist, n_neighbors = 15, min_dist = 0.001, verbose = TRUE)
}
}
\references{
Belkin, M., & Niyogi, P. (2002).
Laplacian eigenmaps and spectral techniques for embedding and clustering.
In \emph{Advances in neural information processing systems}
(pp. 585-591).
\url{http://papers.nips.cc/paper/1961-laplacian-eigenmaps-and-spectral-techniques-for-embedding-and-clustering.pdf}

McInnes, L., & Healey, J. (2018).
UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction
\emph{arXiv preprint} \emph{arXiv}:1802.03426.
\url{https://arxiv.org/abs/1802.03426}

Tang, J., Liu, J., Zhang, M., & Mei, Q. (2016, April).
Visualizing large-scale and high-dimensional data.
In \emph{Proceedings of the 25th International Conference on World Wide Web}
(pp. 287-297).
International World Wide Web Conferences Steering Committee.
\url{https://arxiv.org/abs/1602.00370}
}
